# server/api/gen.go
//go:build go1.22

// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package api

import (
	"bytes"
	"context"
	"encoding/json"
	"fmt"
	"net/http"
)

// Leave defines model for Leave.
type Leave struct {
	Listener string `json:"listener"`
}

// Listen defines model for Listen.
type Listen struct {
	// Listener this tab
	Listener string `json:"listener"`

	// Person this browser
	Person  string `json:"person"`
	Station int    `json:"station"`
}

// Listening defines model for Listening.
type Listening struct {
	Station Station `json:"station"`
	You     Person  `json:"you"`
}

// Now defines model for Now.
type Now struct {
	// Epoch unix seconds all station timelines start from
	Epoch int64 `json:"epoch"`

	// Now unix milliseconds
	Now int64 `json:"now"`

	// Segment seconds per timeline segment
	Segment int `json:"segment"`
}

// Person defines model for Person.
type Person struct {
	Emoji string `json:"emoji"`

	// Name derived from the id a browser keeps
	Name string `json:"name"`
}

// Station defines model for Station.
type Station struct {
	// Id FM frequency in tenths of a MHz
	Id int `json:"id"`

	// LastHeard unix seconds of the last heartbeat
	LastHeard int64 `json:"lastHeard"`

	// Listeners people tuned in
	Listeners int `json:"listeners"`

	// People who is tuned in, in the order they arrived
	People []Person `json:"people"`
}

// StationList defines model for StationList.
type StationList struct {
	// Listeners people listening right now across the band
	Listeners int       `json:"listeners"`
	Stations  []Station `json:"stations"`
}

// LeaveJSONRequestBody defines body for Leave for application/json ContentType.
type LeaveJSONRequestBody = Leave

// ListenJSONRequestBody defines body for Listen for application/json ContentType.
type ListenJSONRequestBody = Listen

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Leave The tab stopped or closed; drop it from the counts right away.
	// (POST /leave)
	Leave(w http.ResponseWriter, r *http.Request)
	// Listen Heartbeat from a tab tuned to a station. Send every 10 seconds while playing; a tab expires after 25.
	// (POST /listen)
	Listen(w http.ResponseWriter, r *http.Request)
	// GetNow Server clock, so every listener computes the same station position.
	// (GET /now)
	GetNow(w http.ResponseWriter, r *http.Request)
	// ListStations Stations that have or recently had listeners.
	// (GET /stations)
	ListStations(w http.ResponseWriter, r *http.Request)

	// (GET /version)
	GetVersion(w http.ResponseWriter, r *http.Request)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// Leave operation middleware
func (siw *ServerInterfaceWrapper) Leave(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.Leave(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// Listen operation middleware
func (siw *ServerInterfaceWrapper) Listen(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.Listen(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetNow operation middleware
func (siw *ServerInterfaceWrapper) GetNow(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetNow(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ListStations operation middleware
func (siw *ServerInterfaceWrapper) ListStations(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListStations(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetVersion operation middleware
func (siw *ServerInterfaceWrapper) GetVersion(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetVersion(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of [http.ServeMux].
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	http.Handler
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/now", wrapper.GetNow)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/stations", wrapper.ListStations)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/listen", wrapper.Listen)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/leave", wrapper.Leave)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/version", wrapper.GetVersion)

	return m
}

type LeaveRequestObject struct {
	Body *LeaveJSONRequestBody
}

type LeaveResponseObject interface {
	VisitLeaveResponse(w http.ResponseWriter) error
}

type Leave204Response struct {
}

func (response Leave204Response) VisitLeaveResponse(w http.ResponseWriter) error {
	w.WriteHeader(204)
	return nil
}

type ListenRequestObject struct {
	Body *ListenJSONRequestBody
}

type ListenResponseObject interface {
	VisitListenResponse(w http.ResponseWriter) error
}

type Listen200JSONResponse Listening

func (response Listen200JSONResponse) VisitListenResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type Listen400Response struct {
}

func (response Listen400Response) VisitListenResponse(w http.ResponseWriter) error {
	w.WriteHeader(400)
	return nil
}

type GetNowRequestObject struct {
}

type GetNowResponseObject interface {
	VisitGetNowResponse(w http.ResponseWriter) error
}

type GetNow200JSONResponse Now

func (response GetNow200JSONResponse) VisitGetNowResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type ListStationsRequestObject struct {
}

type ListStationsResponseObject interface {
	VisitListStationsResponse(w http.ResponseWriter) error
}

type ListStations200JSONResponse StationList

func (response ListStations200JSONResponse) VisitListStationsResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetVersionRequestObject struct {
}

type GetVersionResponseObject interface {
	VisitGetVersionResponse(w http.ResponseWriter) error
}

type GetVersion200JSONResponse struct {
	Version string `json:"version"`
}

func (response GetVersion200JSONResponse) VisitGetVersionResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Leave The tab stopped or closed; drop it from the counts right away.
	// (POST /leave)
	Leave(ctx context.Context, request LeaveRequestObject) (LeaveResponseObject, error)
	// Listen Heartbeat from a tab tuned to a station. Send every 10 seconds while playing; a tab expires after 25.
	// (POST /listen)
	Listen(ctx context.Context, request ListenRequestObject) (ListenResponseObject, error)
	// GetNow Server clock, so every listener computes the same station position.
	// (GET /now)
	GetNow(ctx context.Context, request GetNowRequestObject) (GetNowResponseObject, error)
	// ListStations Stations that have or recently had listeners.
	// (GET /stations)
	ListStations(ctx context.Context, request ListStationsRequestObject) (ListStationsResponseObject, error)

	// (GET /version)
	GetVersion(ctx context.Context, request GetVersionRequestObject) (GetVersionResponseObject, error)
}

type StrictHandlerFunc func(ctx context.Context, w http.ResponseWriter, r *http.Request, request any) (any, error)
type StrictMiddlewareFunc func(f StrictHandlerFunc, operationID string) StrictHandlerFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w http.ResponseWriter, r *http.Request, err error)
	ResponseErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictHTTPServerOptions{
		RequestErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		},
		ResponseErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictHTTPServerOptions) ServerInterface {
	if options.RequestErrorHandlerFunc == nil {
		options.RequestErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	if options.ResponseErrorHandlerFunc == nil {
		options.ResponseErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		}
	}
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictHTTPServerOptions
}

// Leave operation middleware
func (sh *strictHandler) Leave(w http.ResponseWriter, r *http.Request) {
	var request LeaveRequestObject

	var body LeaveJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.Leave(ctx, request.(LeaveRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "Leave")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(LeaveResponseObject); ok {
		if err := validResponse.VisitLeaveResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// Listen operation middleware
func (sh *strictHandler) Listen(w http.ResponseWriter, r *http.Request) {
	var request ListenRequestObject

	var body ListenJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.Listen(ctx, request.(ListenRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "Listen")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(ListenResponseObject); ok {
		if err := validResponse.VisitListenResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetNow operation middleware
func (sh *strictHandler) GetNow(w http.ResponseWriter, r *http.Request) {
	var request GetNowRequestObject

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetNow(ctx, request.(GetNowRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetNow")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetNowResponseObject); ok {
		if err := validResponse.VisitGetNowResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// ListStations operation middleware
func (sh *strictHandler) ListStations(w http.ResponseWriter, r *http.Request) {
	var request ListStationsRequestObject

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.ListStations(ctx, request.(ListStationsRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "ListStations")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(ListStationsResponseObject); ok {
		if err := validResponse.VisitListStationsResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetVersion operation middleware
func (sh *strictHandler) GetVersion(w http.ResponseWriter, r *http.Request) {
	var request GetVersionRequestObject

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetVersion(ctx, request.(GetVersionRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetVersion")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetVersionResponseObject); ok {
		if err := validResponse.VisitGetVersionResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}